requirements.txt
streamlit >= 1.29.0
sentence-transformers >= 2.2.2
pdfplumber >= 0.10.3
python-docx >= 1.1.0
numpy >= 1.21.0
fpdf2 >= 2.7.9
beautifulsoup4 >= 4.11.0
transformers >= 4.21.0
torch >= 2.0.0
accelerate >= 0.20.0
